refactor: Move routes typing to routes folder
fix: Correct the path that broke from moving the routes typing
