+++ src/pages/Login/SigninForm/index.tsx
@@ -11,7 +11,7 @@ import authApi from "@services/authApi";
 import { TouchableOpacity } from "react-native-gesture-handler";
 import TextBox from "@components/TextBox";
 import GenericButton from "@components/GenericButton";
-import { NavigationType } from "@type/routeType";
+import { NavigationType } from "@routes/type";
 import Inputs from "../Inputs";
 
 const SigninForm: React.FC = () => {
