+++ src/pages/Policy/index.tsx
@@ -2,10 +2,12 @@ import { Sc } from "./style";
 import { Gs } from "src/styles/globalStyles";
 import Terms from "./Terms";
 import Entypo from "react-native-vector-icons/Entypo";
-import { NavigationType } from "@type/routeType";
+import { NavigationType } from "@routes/type";
 import { useNavigation } from "@react-navigation/native";
+
 const Policy: React.FC = () => {
   const navigation = useNavigation<NavigationType>();
+
   return (
     <Sc.Container nestedScrollEnabled contentContainerStyle={{ flexGrow: 1 }}>
       <Entypo
@@ -15,6 +17,7 @@ const Policy: React.FC = () => {
         color="#9BABB7"
         onPress={() => navigation.navigate("SignUp")}
       />
+
       <Sc.Wrapper>
         <Gs.Title>{`Termos de Uso e Política de Privacidade`}</Gs.Title>
         <Terms />
