+++ src/pages/ResetPassword/3-NewPassword/SubmitButtons/index.tsx
@@ -2,7 +2,7 @@ import { useNavigation } from "@react-navigation/native";
 import GenericButton from "../../../../components/GenericButton";
 import { SubmitButtonsArray, SubmitButtonsProps } from "../type";
 import React from "react";
-import { NavigationType } from "@type/routeType";
+import { NavigationType } from "@routes/type";
 import { Sc } from "./style";
 
 const SubmitButtons: React.FC<SubmitButtonsProps> = ({ SubmitPassword, isLoading }) => {
