+++ src/pages/SignUp/TermsOfService/index.tsx
@@ -1,7 +1,7 @@
 import TextBox from "@components/TextBox";
 import { Sc } from "./style";
 import { useNavigation } from "@react-navigation/native";
-import { NavigationType } from "@type/routeType";
+import { NavigationType } from "@routes/type";
 
 const TermsOfService = () => {
   const navigation = useNavigation<NavigationType>();
