+++ src/pages/Team/Role/index.tsx
@@ -3,7 +3,6 @@ import React from "react";
 import { Sc } from "./style";
 import { RolesData } from "./type";
 import Integrant from "../Integrant";
-import { Dimensions } from "react-native";
 
 const Role: React.FC<RolesData> = ({ title, integrants }) => {
   return (
